requirements.txt
Flask==3.0.2
imageio==2.34.1
numpy==1.26.4
Pillow==10.3.0
